Complete refactor: Split main.go into focused modules
Split the 929-line main.go into 6 focused modules for better maintainability:

New module structure:
- main.go (39 lines): Entry point, flags, main() only
- config.go (146 lines): Configuration management, SafeConfig, live reload
- text.go (32 lines): Text utilities (formatTime, scrollText)
- artwork.go (248 lines): Image processing, color extraction, Kitty protocol
- model.go (335 lines): Bubble Tea model, Update(), business logic
- view.go (168 lines): View rendering with lipgloss

Benefits:
✓ 95% reduction in main.go size (929 → 39 lines)
✓ Clear separation of concerns
✓ Easier to test individual components
✓ Improved code navigation and maintainability
✓ All tests passing
✓ No functional changes

The refactoring maintains identical functionality while dramatically improving
code organization. Each module now has a single, well-defined responsibility.
